Set media store synchronisation asynchronous.
As syncing media store and shared preferences is longer as the number
of pictures grows, we make the synchronisation asynchronous.
Now, the user has to wait for the synchronisation to end before
being able to watch the upload journal.
A simplification and clarification of the code is still needed, see you soon :-)